+++ packages/roomkit-react/src/Prebuilt/components/RoleChangeRequest/RoleChangeRequestModal.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react';
-import { useMedia } from 'react-use';
 import {
   selectLocalPeerName,
   selectLocalPeerRoleName,
@@ -9,13 +8,13 @@ import {
   useHMSStore,
 } from '@100mslive/react-sdk';
 // @ts-ignore: No implicit Any
-import { PreviewControls, PreviewTile } from './Preview/PreviewJoin';
-import { Box, Button, config as cssConfig, Dialog, Flex, Text } from '../..';
-import { Sheet } from '../../Sheet';
+import { PreviewControls, PreviewTile } from '../Preview/PreviewJoin';
+import { Flex, Text } from '../../..';
 // @ts-ignore: No implicit Any
-import { useMyMetadata } from './hooks/useMetadata';
+import { useMyMetadata } from '../hooks/useMetadata';
 // @ts-ignore: No implicit Any
-import { ROLE_CHANGE_DECLINED } from '../common/constants';
+import { ROLE_CHANGE_DECLINED } from '../../common/constants';
+import { RequestPrompt } from './RequestPrompt';
 
 export const RoleChangeRequestModal = () => {
   const hmsActions = useHMSActions();
@@ -72,7 +71,7 @@ export const RoleChangeRequestModal = () => {
       title={`You're invited to join the ${roleChangeRequest.role.name} role`}
       onOpenChange={async value => {
         if (!value) {
-          await hmsActions.rejectChangeRole(roleChangeRequest);
+          hmsActions.rejectChangeRole(roleChangeRequest);
           sendEvent({ ...roleChangeRequest, peerName: name }, { peerId: roleChangeRequest.requestedBy?.id });
           await hmsActions.cancelMidCallPreview();
           await hmsActions.lowerLocalPeerHand();
@@ -86,66 +85,5 @@ export const RoleChangeRequestModal = () => {
       }}
       actionText="Accept"
     />
-  );
-};
-
-const RequestPrompt = ({
-  open = true,
-  onOpenChange,
-  title,
-  body,
-  actionText = 'Accept',
-  onAction,
-}: {
-  open?: boolean;
-  onOpenChange: (value: boolean) => void;
-  title: string;
-  body: React.ReactNode;
-  actionText?: string;
-  onAction: () => void;
-}) => {
-  const isMobile = useMedia(cssConfig.media.md);
-  const actions = (
-    <Flex justify="center" align="center" css={{ width: '100%', gap: '$md', '@md': { mt: '$8', px: '$8' } }}>
-      <Box css={{ width: '50%' }}>
-        <Dialog.Close css={{ width: '100%' }}>
-          <Button variant="standard" outlined css={{ width: '100%' }}>
-            Decline
-          </Button>
-        </Dialog.Close>
-      </Box>
-      <Box css={{ width: '50%' }}>
-        <Button variant="primary" css={{ width: '100%' }} onClick={onAction}>
-          {actionText}
-        </Button>
-      </Box>
-    </Flex>
-  );
-
-  if (isMobile) {
-    return (
-      <Sheet.Root open={open} onOpenChange={onOpenChange}>
-        <Sheet.Content css={{ py: '$8' }}>
-          <Text css={{ fontWeight: '$semiBold', c: '$on_surface_high', '@md': { px: '$8' } }}>{title}</Text>
-          {body}
-          {actions}
-        </Sheet.Content>
-      </Sheet.Root>
-    );
-  }
-
-  return (
-    <Dialog.Root open={open} onOpenChange={onOpenChange}>
-      <Dialog.Portal>
-        <Dialog.Overlay />
-        <Dialog.Content css={{ p: '$10' }}>
-          <Dialog.Title css={{ p: 0, display: 'flex', flexDirection: 'row', gap: '$md', justifyContent: 'center' }}>
-            <Text variant="h6">{title}</Text>
-          </Dialog.Title>
-          <Box css={{ mt: '$4', mb: '$10' }}>{body}</Box>
-          {actions}
-        </Dialog.Content>
-      </Dialog.Portal>
-    </Dialog.Root>
   );
 };
